+++ sync-for-payroll/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.2.1] sync-for-payroll
 ### Releases
 - [NPM v1.2.1] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.2.1 - sync-for-payroll
+
+## 2023-09-29 00:05:20
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.3.0] sync-for-payroll
+### Releases
+- [NPM v1.3.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.3.0 - sync-for-payroll